Add files via upload
In this game basically, the value to the cards is being assigned and the cards are checked against each other just like it is done in the cards of Pokemon. This basically gives a brief introduction about how to practically apply the classes, into simple application. Game logic is what i have worked upon, one it is quite interesting.